requirements.txt
# Core smolagents installation
smolagents==1.23.0

# Optional extras - install with specific features
smolagents[toolkit]==1.23.0
smolagents[mcp]==1.23.0
smolagents[docker]==1.23.0
smolagents[telemetry]==1.23.0
smolagents[gradio]==1.23.0
smolagents[litellm]==1.23.0

# Phoenix telemetry integration
arize-phoenix
openinference-instrumentation-smolagents

# RAG and multi-agent dependencies
pandas
langchain-core
langchain-community
langchain-text-splitters
sentence-transformers
datasets
rank_bm25

# Web scraping
markdownify
requests

# Environment management
python-dotenv
